Fix VR_* pins & change most passives to 0402
This fixes the power pin wiring as mentioned in the RevB errata:
SX1257 VR_* pins are now only connected to decoupling caps.

This also moves a lot of caps away from the PMOD header & closer to
the appropriate power pins.

Space was getting tight, so this also changes small caps & inductors
to 0402. As a result, there's also a bit of re-routing to reconnect things.
The RF paths should flow a bit better now.